Map assets preparation background processing. Launched in onCreate. Observes readiness to then call getMapAsync
-add background asset preparation via a kotlin coroutine
-expose readiness in activity model via LiveData

building only one asset to validate mechanics

+++ app/src/main/java/com/f8full/feedlocator/data/network/FeedElementBounds.kt
@@ -8,11 +8,11 @@ import com.google.gson.annotations.SerializedName
 class FeedElementBounds {
 
     @SerializedName("min_lat")
-    var minLat : Float? = null
+    var minLat : Double? = null
     @SerializedName("max_lat")
-    var maxLat : Float? = null
+    var maxLat : Double? = null
     @SerializedName("min_lon")
-    var minLon : Float? = null
+    var minLon : Double? = null
     @SerializedName("max_lon")
-    var maxLon : Float? = null
+    var maxLon : Double? = null
 }
